css/css-position/position-sticky-left.html

<!DOCTYPE html>
<title>position:sticky elements should respect the left constraint</title>
<link rel="help" href="https://www.w3.org/TR/css-position-3/#sticky-pos" />
<meta name="assert" content="This test checks that position:sticky elements obey their left anchor after scrolling" />

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
.scroller {
  position: relative;
  width: 200px;
  height: 100px;
  overflow-x: auto;
  overflow-y: hidden;
}

.contents {
  height: 100%;
  width: 500px;
}

.prepadding {
  display: inline-block;
  height: 100%;
  width: 100px;
}

.container {
  display: inline-block;
  height: 100%;
  width: 200px;
}

.innerpadding {
  display: inline-block;
  height: 100%;
  width: 50px;
}

.sticky {
  background-color: green;
  position: sticky;
  left: 50px;
}

.box {
  display: inline-block;
  height: 100%;
  width: 100px;
}

</style>

<div id="scroller1" class="scroller">
  <div class="contents">
    <!-- As these elements are inline, they are whitespace sensitive. -->
    <div class="prepadding"></div><div class="container"><div class="innerpadding"></div><div class="sticky box"></div></div>
  </div>
</div>

<div id="scroller2" class="scroller">
  <div class="contents">
    <!-- As these elements are inline, they are whitespace sensitive. -->
    <div class="prepadding"></div><div class="container"><div class="innerpadding"></div><div class="sticky box"></div></div>
  </div>
</div>

<div id="scroller3" class="scroller">
  <div class="contents">
    <!-- As these elements are inline, they are whitespace sensitive. -->
    <div class="prepadding"></div><div class="container"><div class="innerpadding"></div><div class="sticky box"></div></div>
  </div>
</div>

<script>
test(() => {
  const scroller = document.getElementById('scroller1');
  const prepadding = scroller.querySelector('.prepadding');
  const innerpadding = scroller.querySelector('.innerpadding');
  const sticky = scroller.querySelector('.sticky');

  scroller.scrollLeft = 50;

  const nonStickyLeftY = prepadding.clientWidth + innerpadding.clientWidth;
  assert_equals(sticky.offsetLeft, nonStickyLeftY);
}, 'before reaching the sticking point the sticky box should not be offset');

test(() => {
  const scroller = document.getElementById('scroller2');
  const prepadding = scroller.querySelector('.prepadding');
  const innerpadding = scroller.querySelector('.innerpadding');
  const sticky = scroller.querySelector('.sticky');

  scroller.scrollLeft = 125;

  // This math actually cancels to sticky.offsetLeft == (scroller.scrollLeft + 50),
  // but for clarity the calculations are left explicit.
  const nonStickyLeftY = prepadding.clientWidth + innerpadding.clientWidth;
  const targetLeftY = scroller.scrollLeft + 50;
  const stickyOffset = targetLeftY - nonStickyLeftY;

  assert_equals(sticky.offsetLeft, nonStickyLeftY + stickyOffset);
}, 'after reaching the sticking point the sticky box should be offset');

test(() => {
  const scroller = document.getElementById('scroller3');
  const container = scroller.querySelector('.container');
  const prepadding = scroller.querySelector('.prepadding');
  const innerpadding = scroller.querySelector('.innerpadding');
  const sticky = container.querySelector('.sticky');

  scroller.scrollLeft = 200;

  const nonStickyLeftY = prepadding.clientWidth + innerpadding.clientWidth;
  const maxOffsetInContainer = container.clientWidth - innerpadding.clientWidth
      - sticky.clientWidth;

  assert_equals(sticky.offsetLeft, nonStickyLeftY + maxOffsetInContainer);
}, 'the sticky box should not be pushed outside its containing block');
</script>
